+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -224,6 +224,7 @@ function gwoCard() {
       ];
       const expandedCards = [
         "gwaio_enable_planetaryradar",
+        "gwaio_health_titans",
         "gwaio_speed_structure",
         "gwaio_upgrade_advancedairfactory",
         "gwaio_upgrade_advancedbotfactory",