+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -4,6 +4,7 @@ define([
 ], function (gwoUnit, gwoGroup) {
   const vehicleFactories = [gwoUnit.vehicleFactory, gwoUnit.vehicleFactoryAdvanced]     
   const botFactories = [gwoUnit.botFactory, gwoUnit.botFactoryAdvanced]
+  const airFactories = [gwoUnit.airFactory, gwoUnit.airFactoryAdvanced]
   const mobileStructures = gwoGroup.structuresDefences.concat(
     gwoGroup.structuresIntel,
     gwoGroup.structuresArtillery,
@@ -435,6 +436,7 @@ define([
       { id: "gwaio_combat_titans", units: gwoGroup.titans },
       { id: "gwaio_cooldown_vehicles", units: vehicleFactories },
       { id: "gwaio_cooldown_bots", units: botFactories },
+      { id: "gwaio_cooldown_air", units: airFactories },
       { id: "gwaio_speed_structure", units: mobileStructures },
     ],
   };