+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_start/setup.js
@@ -380,7 +380,7 @@ function gwoSetup() {
           const busyToken = {};
           model.makeGameBusy(busyToken);
 
-          const version = "5.66.2";
+          const version = "5.67.0";
           console.log("War created using Galactic War Overhaul v" + version);
 
           const game = new GW.Game();