+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## Unreleased
+
+### Added
+
+- New tech cards
+  - Titan Armour Tech
+
 ## v5.66.2 - 2024-07-28
 
 ### Fixed